cleaning_service/requirements.txt
fastapi
uvicorn[standard]
pydantic<2
aio-pika
python-dotenv